+++ src/suites/baby_jubjub.rs
@@ -105,6 +105,7 @@ pub(crate) mod tests {
         const SUITE_NAME: &str = "baby-jubjub_sha-512_tai";
     }
 
+    suite_tests!(ThisSuite);
     tiny_suite_tests!(ThisSuite);
     pedersen_suite_tests!(ThisSuite);
     thin_suite_tests!(ThisSuite);